+++ comply54/autogen/__init__.py
@@ -1,0 +1,3 @@
+from .adapter import Comply54UserProxy, register_compliance, register_compliance_guard
+
+__all__ = ["Comply54UserProxy", "register_compliance", "register_compliance_guard"]